Bring back `describe` sub-command
When you only have a single sub-command, `validate` in this case,
Typer seems to automatically move that functionality to the top-level
command. To get around that strangeness, I've re-added the `describe`
command that's intended to print out the file format and the validation
list, but isn't implemented yet. For now if you call it, it simply
reports _Feature coming soon..._

We'll implement this for real in a follow-up PR sometime soon.