feat: Replace `main.py` with Typer-based CLI app (#63)
### Replace `main.py` to [Typer](https://typer.tiangolo.com/) CLI app

The CLI setup in `main.py` is pretty primitive, and is really geared
towards _us_, developers on this project. I would like to create a
utility that outside technical users can use to test their files against
our validation logic. To get there without having to reinvent many
wheels, we need an actual CLI framework. That's where
[Typer](https://typer.tiangolo.com/) comes in.

Typer is a wrapper around
[Click](https://click.palletsprojects.com/en/8.1.x/), written by the
same dev as FastAPI, so it follows many of the same patterns as we're
used to there, such as leaning into Python's `typing` and
[Pydantic](https://docs.pydantic.dev/latest/) for example.


### Return validation results as Pandas `DataFrame` instead of vanilla
`dict`

I also wanted to support multiple output formats; at least json, csv,
and default to a human-readable table. And while we _could_ do that with
a `dict`, it's much easier to work with `DataFrame`s as they're already
tabular in nature and have a ton of transform functionality built-in.
So, I have also refactored the `validate` function to return a Panda's
`DataFrame` instead of a vanilla `dict`. This made it much easier to
output validation data in multiple formats.

Along with this, I did a fair bit of refactoring of that surrounding
code, breaking up the big monolith function into smaller testable
functions.

### `cfpb-val` script

To simply using the CLI, I've added a shortcut script so you don't have
to call the full path to the Python script. In it's current form, you
can now call just...

    cfpb-val --help

...instead of...

    python regtech_data_validator/cli.py --help 

#### Better script name?

I'm not in love with `cfpb-val`. I started off with `cfpb-comply`, but I
didn't love that either. I'm looking for something that's short, but
also very clear that it's CFPB-related, and having to do with this
project. `cfpb-rdv`?

### Lots of `README.md` love

I've reworked the `README.md` quite a bit, focusing on users coming to
this repo for the first time, and trying to get the CLI to work. This
includes pushing the Poetry-based install up to the top and simplifying
it, slimming and pushing down development-related setup, and adding
details about how to use the CLI itself.

See:
https://github.com/cfpb/regtech-data-validator/blob/add-typer-cli/README.md

### Refactor `lei` arg to more generic `context`

In attempt to future-proof this a bit, I've added a multi-value
`context` arg to the CLI, which allows you to pass arbitrary
`<key>=<value>` pairs, `lei` being the only key that's current used for
anything.

### Other tidbits I found along the way

- Removed deprecated `python.formatting.provider` VS Code configuration
- Removes no longer used `Makefile`
- Fixes several validations that had incorrect IDs, severities, or were
in the wrong phase

---------

Co-authored-by: lchen-2101 <73617864+lchen-2101@users.noreply.github.com>