Adds the test_csv_differences.py to check for discrepancies between the python code and the csv located at https://raw.githubusercontent.com/cfpb/sbl-content/main/fig-files/validation-spec/2024-validations.csv
There was no automated check to verify that the code and csv were in sync.  Because changes to the csv had been made
but not flowed into the code, several differences were found.  This will make it easier to identify changes and implement them
in the code or correct them in the csv.

https://raw.githubusercontent.com/cfpb/sbl-content/main/fig-files/validation-spec/2024-validations.csv