68 correct validation description errors (#70)
These are all changes to make our python code consistent with the CVS
located at
https://raw.githubusercontent.com/cfpb/sbl-content/main/fig-files/validation-spec/2024-validations.csv

This goes hand in hand with issue #65 

Of note is the addition of W0680, which checks the Census GEO id against
the census CSV.

Majority of the changes involved missing spaces, copy/paste errors, etc.

The updates to this file are still in progress as there are 4 errors
(E0940, E1120, E1300, E1480) that have descriptions that Chynna and
Shomari are looking into to verify if the CSV is correct. So this PR is
simply to get a head start.