65 create test to make sure validations stay in sync with 2024 validationscsv (#69)
This adds a pytest (test_csv_differences.py) to validate our python code
against the CSV located at
https://raw.githubusercontent.com/cfpb/sbl-content/main/fig-files/validation-spec/2024-validations.csv

This will compare error/warning codes (making sure neither the code nor
csv have codes the other doesn't), the type (error or warning) and the
description.

Special note is taken of E2014 and E2015 due to formatting in the CSV.
In the near future when the frontend is ready to start displaying
error/warning descriptions, discussions will be had to figure out how we
want to display the more complicated descriptions and what sort of
formatting the backend should have. Right now, we preserve as much of
the formatting as we can but the pytest will also strip all of this off
for these two errors (or any others added to the remove_formatting list)
and compare just character data. In general, we do NOT want to do this
because several strings in our python code were missing spaces and other
standard grammatical formatting, and stripping that off would have
caused the test to improperly accept that description.

This story is being worked in conjunction with #68 which is being used
to update the phase_validations.py for other discrepancies found during
testing. It is being routinely merged into this branch to properly run
the pytest.