Added versions to the ruff and black linter actions (#90)
Closes #89 

Added versions to the linter actions. Since the pyproject.toml specifies
distinct versions for each (instead of say ^24.1.0), I put the version
equal to what is in the toml. Note that, as in the other stories
referenced, the black linter supports the concept of capability
versions. We're just not using itwith:
          version: 0.0.278here.