122 add fig anchor metadata to validations (#124)
Closes #122 

I created a script that read through the fig_source.html, which I saved
from browser, and found each of the elements containing the validation
ID for each SBLCheck. Then I grabbed the href and stored that in the
fig_anchor in the phase_validations.py file.

I created a pytest to use the same html to loop through both the phase
validation fig_anchors and the hrefs to compare to ensure each
validation ID had the correct fig anchor. I also updated the existing
test_cli.py formats to include the new fig_anchor field.

Might be a good idea to store this tool somewhere, it's all local
currently. Just in case they change something, it makes it easy to loop
through the py file and insert the fig_anchor instead of manually
copy/pasting.

Updated the schema to include the fig_anchor in the check, which
automatically carries over into the validation results.
Updated the df_to_json to include the fig_anchor in the json we send to
clients. The other df_to's automatically get the fig_anchor.

I don't think we need another story in the filing-api, the fig_anchor
for each result will be in the JSON blob.

I would LOVE to figure out how to automatically pull down the FIG html
in an actual usable way for the pytests. However, because it's all
javascripted, the actual hrefs don't come across as full links if you do
a request.get, or curl, or wget. The only way I've found to have both
the full href link and the Validation ID associated with it is saving
the page off in a browser.