Updated df_to_download to create an empty csv with headers (#136)
Closes #135 

I did this so the filing-api will always have a report to upload. There
will be no confusion if someone looks in the S3, or tries to download a
report. The report will always be there, it's just a successfully
validated submission that has no errors or warning will be empty
(essentially like an empty json being stored in the submission object).

If we don't like this approach, I'll need to update the /report endpoint
story to only call df_to_download if there are errors/warnings, and then
on that /report endpoint GET, check that a report exists in the S3 first
before trying to read it. Right now the assumption is a report will
always be there, even if empty.