Added comparison_helper that uses python eval to check value comaprisons (#152)
Closes #146 

Turns out our is_less_than, is_greater_than and
is_greater_than_or_equals all had the same issue. So I created a
comparison_helper function that uses python eval() to convert the value
to its type (so we're not beholden to int, float, etc in the functions)
and then uses eval again with the passed in comparison sign to evaluate
the value and the limit.

Also beefed up the pytests for all three functions to use values that
would actually cause the error. Simple strings like "1" and "2" won't
cut it.