151 change formatters and groupby fields to maintain order (#154)
Closes #151 

Updated areas of code called out in the story.

Attached a before and after csv to show what was happening. We don't
have any pytests that are this verbose to find those errors/warnings
that were causing the reordering. Can add those if desired.

[logical_errors_example.csv](https://github.com/cfpb/regtech-data-validator/files/14961898/logical_errors_example.csv)

[old_order_logical_errors_example.csv](https://github.com/cfpb/regtech-data-validator/files/14961899/old_order_logical_errors_example.csv)

Notice the ordering of the fieds for E2014 and W2035. The old format had
them reordered lexicographically, whereas the new order maintains what
is in the FIG and in the groupby order of the Checks.